MA-15: added ability to publish post on main page
