+++ src/Redux/State.js
@@ -1,3 +1,5 @@
+import { rerenderEntireTree } from '../render'
+
 let state = {
 	profilePage: {
 		posts: [
@@ -28,8 +30,8 @@ export let addPost = postMessage => {
 		message: postMessage,
 		likesCounter: 0,
 	}
-
 	state.profilePage.posts.push(newPost)
+	rerenderEntireTree(state)
 }
 
 export default state
