+++ src/index.js
@@ -1,22 +1,5 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
 import './index.css'
-import reportWebVitals from './reportWebVitals'
-import App from './App'
 import state from './Redux/State'
-import { addPost } from './Redux/State'
-import { BrowserRouter } from 'react-router-dom'
+import { rerenderEntireTree } from './render'
 
-const root = ReactDOM.createRoot(document.getElementById('root'))
-
-root.render(
-	<React.StrictMode>
-		<BrowserRouter>
-			<App state={state} addPost={addPost} />
-		</BrowserRouter>
-	</React.StrictMode>
-)
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals()
+rerenderEntireTree(state)
